requirements.txt
# Google API
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

python-dotenv==1.0.1
openai==1.60.2
docx2txt==0.8
unstructured

# Langchain, Langgraph, GenAI
langchain==0.3.19
langchain-chroma==0.2.2
langchain-huggingface==0.1.2
langchain_community==0.3.17
langchain-openai==0.3.6
langgraph==0.2.73 
langgraph-checkpoint-sqlite==2.0.5
google-generativeai==0.8.4
pytest-playwright==0.7.0
playwright==1.50.0
